refactor(cron): add support to multiples networks in cron
