+++ src/pkg/cron/cron.ts
@@ -1,10 +1,11 @@
 import { IDatabase } from "../database/type";
 import { Contract } from "ethers";
 import { toTransaction } from "./util";
-import { Transaction, TransactionType } from "../../sychronizer";
-import { ERROR_MSG_NOT_INITIALIZED } from "./type";
+import { Network, Transaction, TransactionType } from "../../synchronizer";
+import { ERROR_MSG_NOT_INITIALIZED, ICron } from "./type";
 
-export class Cron {
+export class Cron implements ICron {
+  private network: Network;
   private intervalMs: number;
   private db: IDatabase;
   private contract: Contract;
@@ -13,7 +14,8 @@ export class Cron {
   private latestWithdrawBlock: number;
   private ready: boolean;
 
-  constructor(intervalMs: number, db: IDatabase, contract: Contract) {
+  constructor(network: Network, intervalMs: number, db: IDatabase, contract: Contract) {
+    this.network = network;
     this.intervalMs = intervalMs;
     this.db = db;
     this.contract = contract;
@@ -26,7 +28,7 @@ export class Cron {
     // get latest deposit transaction from db
     let lastDepositTransaction;
     try {
-      lastDepositTransaction = await this.db.getLastTransactionByType(TransactionType.Deposit);
+      lastDepositTransaction = await this.db.getLastTransactionByType(this.network, TransactionType.Deposit);
       this.latestDepositBlock = lastDepositTransaction.block;
     } catch (err: any) {
       console.warn(err.message);
@@ -35,7 +37,7 @@ export class Cron {
     // get latest withdraw transaction from db
     let lastWithdrawTransaction;
     try {
-      lastWithdrawTransaction = await this.db.getLastTransactionByType(TransactionType.Withdraw);
+      lastWithdrawTransaction = await this.db.getLastTransactionByType(this.network, TransactionType.Withdraw);
       this.latestWithdrawBlock = lastWithdrawTransaction.block;
     } catch (err: any) {
       console.warn(err.message);
@@ -64,19 +66,19 @@ export class Cron {
     // sort and find fetched events hashes inside the database
     const sortedTransactions = events.sort((a, b) => a.blockNumber - b.blockNumber);
     const hashes = sortedTransactions.map((e) => e.transactionHash);
-    const dbHashes = (await this.db.listTransactionsByHashes(hashes)).map((t) => t.hash);
+    const dbHashes = (await this.db.listTransactionsByHashes(this.network, hashes)).map((t) => t.hash);
 
     // filter, parse and prepare transactions to bulk insert in database
     const filtered = sortedTransactions.filter((t) => !dbHashes.includes(t.transactionHash));
     if (filtered.length === 0) {
       return [];
     }
-    const parsed = filtered.map((e) => toTransaction(e, type));
+    const parsed = filtered.map((e) => toTransaction(this.network, e, type));
 
     return parsed;
   }
 
-  public async run() {
+  public async run(): Promise<void> {
     // initialize cronjob instance getting latests deposit/withdraw values
     await this.init();
 
@@ -85,7 +87,7 @@ export class Cron {
       try {
         const deposits = await this.getNewTransactions(TransactionType.Deposit, this.latestDepositBlock);
         if (deposits.length > 0) {
-          console.log("inserting depostir in db", deposits);
+          console.log("inserting deposits in db", deposits);
           await this.db.insertTransactions(deposits);
           this.latestDepositBlock = deposits[deposits.length - 1].block;
         }
